Move more early buffered lints to dyn lint diagnostics (4/N)
